# fct-frontend/api-spec/openapi.yaml
openapi: 3.1.1
info:
    title: REST API for the website of "Academia Fabiana Perez"
    description: REST API for the website of "Academia Fabiana Perez"
    version: 1.0.0
servers:
    - url: http://localhost:8080/api/v1
paths:
...
    /appointments:
        get:
            summary: Get all appointments
            operationId: getAllAppointments
            tags:
                - Appointments
            parameters:
                - name: userId
                  in: query
                  description: Filter by user id
                  required: false
                  schema:
                      type: integer
                      format: int64
                - name: status
                  in: query
                  description: Filter by status
                  required: false
                  schema:
                      $ref: "#/components/schemas/AppointmentStatusEnum"
            responses:
                200:
                    description: Successful operation
                    content:
                        application/json:
                            schema:
                                type: array
                                items:
                                    $ref: "#/components/schemas/Appointment"
        post:
            summary: Add a new appointment
            operationId: addAppointment
            tags:
                - Appointments
            requestBody:
                description: Appointment to add
                required: true
                content:
                    application/json:
                        schema:
                            $ref: "#/components/schemas/AddAppointmentRequest"
            responses:
                201:
                    description: Successful operation
                    content:
                        application/json:
                            schema:
                                $ref: "#/components/schemas/Appointment"
                400:
                    description: Bad request
                    content:
                        application/json:
                            schema:
                                $ref: "#/components/schemas/ValidationErrorMessage"
                401:
                    description: Unauthorized
                    content:
                        application/json:
                            schema:
                                $ref: "#/components/schemas/ErrorMessage"
                403:
                    description: Forbidden
                    content:
                        application/json:
                            schema:
                                $ref: "#/components/schemas/ErrorMessage"
                409:
                    description: Appointment already exists
                    content:
                        application/json:
                            schema:
                                $ref: "#/components/schemas/ErrorMessage"
    /appointments/{id}:
        get:
            summary: Get an appointment by id
            operationId: getAppointmentById
            tags:
                - Appointments
            parameters:
                - name: id
                  in: path
                  description: Appointment id
                  required: true
                  schema:
                      type: integer
                      format: int64
            responses:
                200:
                    description: Successful operation
                    content:
                        application/json:
                            schema:
                                $ref: "#/components/schemas/Appointment"
                400:
                    description: Bad request
                    content:
                        application/json:
                            schema:
                                $ref: "#/components/schemas/ErrorMessage"
                401:
                    description: Unauthorized
                    content:
                        application/json:
                            schema:
                                $ref: "#/components/schemas/ErrorMessage"
                403:
                    description: Forbidden
                    content:
                        application/json:
                            schema:
                                $ref: "#/components/schemas/ErrorMessage"
                404:
                    description: Appointment not found
                    content:
                        application/json:
                            schema:
                                $ref: "#/components/schemas/ErrorMessage"
        put:
            summary: Update an appointment
            operationId: updateAppointment
            tags:
                - Appointments
            parameters:
                - name: id
                  in: path
                  description: Appointment id
                  required: true
                  schema:
                      type: integer
                      format: int64
            requestBody:
                description: Appointment to update
                required: true
                content:
                    application/json:
                        schema:
                            $ref: "#/components/schemas/UpdateAppointmentRequest"
            responses:
                200:
                    description: Successful operation
                    content:
                        application/json:
                            schema:
                                $ref: "#/components/schemas/Appointment"
                400:
                    description: Bad request
                    content:
                        application/json:
                            schema:
                                $ref: "#/components/schemas/ValidationErrorMessage"
                401:
                    description: Unauthorized
                    content:
                        application/json:
                            schema:
                                $ref: "#/components/schemas/ErrorMessage"
                403:
                    description: Forbidden
                    content:
                        application/json:
                            schema:
                                $ref: "#/components/schemas/ErrorMessage"
                404:
                    description: Appointment not found
                    content:
                        application/json:
                            schema:
                                $ref: "#/components/schemas/ErrorMessage"
        delete:
            summary: Delete an appointment
            operationId: deleteAppointment
            tags:
                - Appointments
            parameters:
                - name: id
                  in: path
                  description: Appointment id
                  required: true
                  schema:
                      type: integer
                      format: int64
            responses:
                204:
                    description: Successful operation
                400:
                    description: Bad request
                    content:
                        application/json:
                            schema:
                                $ref: "#/components/schemas/ErrorMessage"
                401:
                    description: Unauthorized
                    content:
                        application/json:
                            schema:
                                $ref: "#/components/schemas/ErrorMessage"
                403:
                    description: Forbidden
                    content:
                        application/json:
                            schema:
                                $ref: "#/components/schemas/ErrorMessage"
                404:
                    description: Appointment not found
                    content:
                        application/json:
                            schema:
                                $ref: "#/components/schemas/ErrorMessage"
        patch:
            summary: Update an appointments status
            operationId: updateAppointmentStatus
            tags:
                - Appointments
            parameters:
                - name: id
                  in: path
                  description: Appointment id
                  required: true
                  schema:
                      type: integer
                      format: int64
            requestBody:
                description: Appointment status
                required: true
                content:
                    application/json:
                        schema:
                            $ref: "#/components/schemas/UpdateAppointmentStatusRequest"
            responses:
                200:
                    description: Appointment updated
                    content:
                        application/json:
                            schema:
                                $ref: "#/components/schemas/Appointment"
                400:
                    description: Bad request
                    content:
                        application/json:
                            schema:
                                $ref: "#/components/schemas/ErrorMessage"
                401:
                    description: Unauthorized
                    content:
                        application/json:
                            schema:
                                $ref: "#/components/schemas/ErrorMessage"
                403:
                    description: Forbidden
                    content:
                        application/json:
                            schema:
                                $ref: "#/components/schemas/ErrorMessage"
                404:
                    description: Appointment not found
                    content:
                        application/json:
                            schema:
                                $ref: "#/components/schemas/ErrorMessage"
...
    /carts:
        get:
            summary: Get all carts
            operationId: getAllCarts
            tags:
                - Carts
            parameters:
                - name: userId
                  in: query
                  description: Filter by user id
                  required: false
                  schema:
                      type: integer
                      format: int64
            responses:
                200:
                    description: Successful operation
                    content:
                        application/json:
                            schema:
                                type: array
                                items:
                                    $ref: "#/components/schemas/Cart"
                400:
                    description: Invalid status value
                    content:
                        application/json:
                            schema:
                                $ref: "#/components/schemas/ErrorMessage"

        post:
            summary: Add Product to Cart
            operationId: addCart
            tags:
                - Carts
            parameters:
                - name: userId
                  in: query
                  description: User to add the product to the cart
                  required: true
                  schema:
                      type: integer
                      format: int64
                - name: productId
                  in: query
                  description: Product to add to the cart
                  required: true
                  schema:
                      type: integer
                      format: int64
            responses:
                201:
                    description: Successful operation
                    content:
                        application/json:
                            schema:
                                $ref: "#/components/schemas/Cart"
                400:
                    description: Invalid status value
                    content:
                        application/json:
                            schema:
                                $ref: "#/components/schemas/ErrorMessage"
                401:
                    description: Unauthorized
                    content:
                        application/json:
                            schema:
                                $ref: "#/components/schemas/ErrorMessage"
                403:
                    description: Forbidden
                    content:
                        application/json:
                            schema:
                                $ref: "#/components/schemas/ErrorMessage"
        put:
            summary: Update Product's amount in a Cart
            operationId: updateCart
            tags:
                - Carts
            parameters:
                - name: userId
                  in: query
                  description: User id
                  required: true
                  schema:
                      type: integer
                      format: int64
                - name: productId
                  in: query
                  description: Product id
                  required: true
                  schema:
                      type: integer
                      format: int64
                - name: quantity
                  in: query
                  description: Product quantity
                  required: true
                  schema:
                      type: integer
                      format: int32
                      minimum: 1
            responses:
                200:
                    description: Successful operation
                    content:
                        application/json:
                            schema:
                                $ref: "#/components/schemas/Cart"
                400:
                    description: Bad request
                    content:
                        application/json:
                            schema:
                                $ref: "#/components/schemas/ErrorMessage"
                401:
                    description: Unauthorized
                    content:
                        application/json:
                            schema:
                                $ref: "#/components/schemas/ErrorMessage"
                403:
                    description: Forbidden
                    content:
                        application/json:
                            schema:
                                $ref: "#/components/schemas/ErrorMessage"
        delete:
            summary: Remove Product from Cart
            operationId: removeProductFromCart
            tags:
                - Carts
            parameters:
                - name: userId
                  in: query
                  description: User id
                  required: true
                  schema:
                      type: integer
                      format: int64
                - name: productId
                  in: query
                  description: Product id
                  required: true
                  schema:
                      type: integer
                      format: int64
            responses:
                204:
                    description: Successful operation
                400:
                    description: Bad request
                    content:
                        application/json:
                            schema:
                                $ref: "#/components/schemas/ErrorMessage"
                401:
                    description: Unauthorized
                    content:
                        application/json:
                            schema:
                                $ref: "#/components/schemas/ErrorMessage"
                403:
                    description: Forbidden
                    content:
                        application/json:
                            schema:
                                $ref: "#/components/schemas/ErrorMessage"
...
components:
    schemas:
...
        AppointmentStatusEnum:
            type: string
            enum: [WAITING, ACCEPTED, DECLINED]
            x-field-extra-annotation: "@ValidEnum(AppointmentStatusEnum.class)"
...
        Appointment:
            type: object
            properties:
                id:
                    type: integer
                    format: int64
                date:
                    type: string
                    format: date
                duration:
                    type: integer
                    format: int32
                status:
                    $ref: "#/components/schemas/AppointmentStatusEnum"
                description:
                    type: string
                price:
                    type: number
                    format: double
                userId:
                    type: integer
                    format: int64
                categoryId:
                    type: integer
                    format: int64
            required:
                [
                    id,
                    date,
                    duration,
                    status,
                    description,
                    price,
                    userId,
                    categoryId,
                ]
...
        Cart:
            type: object
            properties:
                userId:
                    type: integer
                    format: int64
                productId:
                    type: integer
                    format: int64
                quantity:
                    type: integer
                    format: int32
            required: [id, userId, productId, quantity]
...
        AddAppointmentRequest:
            type: object
            properties:
                description:
                    type: string
                    x-field-extra-annotation: '@NotBlank(message = "Description field cannot be blank")'
                date:
                    type: string
                    format: date
                    x-field-extra-annotation: >-
                        @Future(message = "Date must be in the future")
                        @DateTimeFormat(iso = DateTimeFormat.ISO.DATE_TIME)
                duration:
                    type: integer
                    format: int32
                    x-field-extra-annotation: >-
                        @Min(value = 1, message = "Duration must be greater than 0")
                userId:
                    type: integer
                    format: int64
                categoryId:
                    type: integer
                    format: int64
            required: [description, duration, categoryId, date, userId]
        UpdateAppointmentRequest:
            type: object
            properties:
                name:
                    type: string
                description:
                    type: string
                duration:
                    type: integer
                    format: int32
                    x-field-extra-annotation: >-
                        @Min(value = 1, message = "Duration must be greater than 0")
                        @Max(value = 7, message = "Duration must be less than 7")
                date:
                    type: string
                    format: date
                    x-field-extra-annotation: >-
                        @Future(message = "Date must be in the future")
                        @DateTimeFormat(iso = DateTimeFormat.ISO.DATE_TIME)
                categoryId:
                    type: integer
                    format: int64
        UpdateAppointmentStatusRequest:
            type: object
            properties:
                status:
                    $ref: "#/components/schemas/AppointmentStatusEnum"
            required: [status]

        ## ---- AppointmentCategory ----
...
        ErrorMessage:
            type: object
            properties:
                timestamp:
                    type: string
                    format: date-time
                    example: "1970-01-01T00:00+18:00"
                status:
                    type: integer
                    format: int32
                    example: 404
                error:
                    type: string
                    example: "Not Found"
                message:
                    type: string
                    example: "Resource not found"
                path:
                    type: string
                    example: "/api/v1/resource/1"
            required: [timestamp, status, error, message, path]
        ValidationErrorMessage:
            type: object
            properties:
                timestamp:
                    type: string
                    format: date-time
                    example: "1970-01-01T00:00+18:00"
                status:
                    type: integer
                    format: int32
                    example: 404
                error:
                    type: string
                    example: "Bad Request"
                messages:
                    type: object
                    additionalProperties:
                        type: string
                    example:
                        "name": "Name field cannot be blank"
                path:
                    type: string
                    example: "/api/v1/resource/1"
            required: [timestamp, status, error, messages, path]
    securitySchemes:
        BearerAuth:
            description: Bearer token using a JWT
            type: http
            scheme: bearer
            bearerFormat: JWT